Scripts: Refactor to extract license logic (#66179)
Refactor some of the check-license logic into a reusable module.

The check-license script continues to act as the CLI, while another module
contains the license checking logic.

This is helpful for work migrating to npm workspaces:
https://github.com/WordPress/gutenberg/pull/65681

The license checking logic is required, but the implementation will be different
to account for npm workspaces in Gutenberg.

---
