+++ packages/scripts/CHANGELOG.md
@@ -2,12 +2,15 @@
 
 ## Unreleased
 
+### Internal
+
+-   Refactor to extract license related logic to a reusable module ([#66179](https://github.com/WordPress/gutenberg/pull/66179)).
+
 ## 30.2.0 (2024-10-16)
 
 ## 30.1.0 (2024-10-03)
 
 ## 30.0.0 (2024-09-19)
-
 
 ### Breaking Changes
