+++ packages/scripts/utils/test/licenses/bsd3clause.txt
