+++ packages/scripts/utils/test/licenses/mit.txt
